Project_14_RAG_VIBE_CODING/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
chromadb
python-multipart
PyPDF2
pandas
python-dotenv
streamlit
